Unit-testing plugs
Adds tests for our `Auth` plug to make sure that it works in isolation.
We make use of the `bypass_through` test helper in order to avoid
reimplementing our the whole endpoint and router pipeline in order to
test the `authenticate_user` function.

Also eases up the number of password hashing rounds we do in the `test`
environment to speed up our test suite.