Creating slugs
Adds a `videos.slug` column using an Ecto migration. Tweaks the video
changeset to generate the slug automatically for us when creating or
editing a video. Introduces a new protocol to seamlessly build URLs from
those new slugs. Extends the video schema with a custom Ecto type, which
allows Phoenix to interpret the new URL format correctly and get the ID.