Phoenix clients with ES6
Builds a bare-bones client to simply establish a connection to our
channel later. Adds a `Video` object in our JavaScript code, with an
`init` function to set-up the player and pluck our video ID from the
element attributes. Starts the socket connection and initialise our
player with the DOM variables that we will need.