Persisting annotations
Creates the `Multimedia.Annotation` model, which belongs to a video and
a user. Introduces context functions to create an annotation and list
all annotations for a video.

Updates the `VideoChannel` to ensure that all incoming events have the
current user, and that new-annotation events persist to the database.

Adds a JSON view for `User` and `Annotation`, then uses them to pass a
list of any existing annotations to the client when a user joins the
channel. Makes use of the `render_many` and `render_one` functions.